Move deep freeze and ref check functions to a separate file
